Reworked DataContexts
Multiple contexts does not appear possible within WASM-SQLite.
