+++ Columbus.Welkom/Client/Program.cs
@@ -1,5 +1,7 @@
 using Blazored.LocalStorage;
 using Columbus.Welkom.Client;
+using Columbus.Welkom.Client.Repositories.Interfaces;
+using Columbus.Welkom.Client.Repositories;
 using Columbus.Welkom.Client.Services;
 using Columbus.Welkom.Client.Services.Interfaces;
 using KristofferStrube.Blazor.FileSystemAccess;
@@ -8,9 +10,6 @@ using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using SqliteWasmHelper;
 using Radzen;
 using Microsoft.EntityFrameworkCore;
-using Columbus.Welkom.Client.DataContext;
-using Columbus.Welkom.Client.Repositories.Interfaces;
-using Columbus.Welkom.Client.Repositories;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
 builder.RootComponents.Add<App>("#app");
@@ -25,15 +24,7 @@ builder.Services.AddScoped<TooltipService>();
 builder.Services.AddScoped<ContextMenuService>();
 
 // DataContext
-builder.Services.AddSqliteWasmDbContextFactory<PigeonContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<OwnerContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<PigeonRaceContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<RaceContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<LeagueContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<PigeonSwapContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<SelectedYearPigeonContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<SelectedYoungPigeonContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
-builder.Services.AddSqliteWasmDbContextFactory<TeamContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
+builder.Services.AddSqliteWasmDbContextFactory<DataContext>(opts => opts.UseSqlite("Data Source=welkom.sqlite3"));
 
 // Repositories
 builder.Services.AddScoped<IOwnerRepository, OwnerRepository>();
