+++ Columbus.Welkom/Client/Repositories/Interfaces/IPigeonRepository.cs
@@ -5,5 +5,6 @@ namespace Columbus.Welkom.Client.Repositories.Interfaces
 {
     public interface IPigeonRepository : IBaseRepository<PigeonEntity>
     {
+        Task<IEnumerable<PigeonEntity>> GetPigeonsByCountriesAndYearsAndRingNumbers(IEnumerable<Pigeon> pigeons);
     }
 }
